requirements.txt
Django==5.2.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.1.0
